Module surreal_model_layer::model::surreal_store::base_crud

Isomorphic(?) functions the Model Controllers can call

Traits

SurrealBmc

Model Bmc Structs need to implement this trait.

Functions

base_create

Create the data in content in the TABLE of the calling model (Set by the impl of SurrealBmc in the parent calling model). Uses a standard random id.

base_create_with_id

Same as base_create, but uses the given string as the id.

base_create_with_ulid

Same as base_create, but uses a ULID as the id

base_delete

Delete a single record based on the given id

base_get

Get a single record based on the given id

base_info

base_list

List all records

base_update

Update a single record based on the given id with the given data